Added classes for games, difficulties, and game list
An interface was created for game difficulties to enforce typing. A game
class was created to enforce typing (contains default Difficult
defintions). A game-list abstract class ws created, and a game-list-all
service was added inheriting that abstract class. Games are defined in
the static games array in the game-list-all service.

The main-menu, and leaderboards components were updated to reflect the
above changes.